refactor: 부적절한 패키지 구조 변경
